Add files via upload
This code is used to train and test the model for classifying motor impairment rehabilitation states. Due to data privacy concerns and the patients' own consent, only a subset of patients' plantar pressure data is provided. However, detailed descriptions of the device setup and experimental conditions required for full reproduction can be found in the manuscript and supplementary information.